create a guild info base, further utilize init, fix some issues with message commands, add separator items for eval dict

+++ commands/bot_owner/messages/delete.py
@@ -2,6 +2,7 @@ import discord
 from discord.abc import Messageable
 
 from bot import Interaction
+from commands.bot_owner._base import TextChannelTypes
 from core.exceptions import send_bad_argument, send_unknown_error
 
 # ⸻⸻⸻⸻⸻⸻⸻⸻⸻⸻⸻⸻⸻⸻
@@ -10,20 +11,26 @@ from core.exceptions import send_bad_argument, send_unknown_error
 
 async def run_bo_messages_delete(
     interaction : Interaction,
-    channel     : Messageable,
     message_id  : str,
+    channel     : Messageable | None = None,
 ) -> None:
     _ = await interaction.response.defer(ephemeral = True)
 
+    target_channel = channel or interaction.channel
+
+    if not isinstance(target_channel, TextChannelTypes):
+        await send_bad_argument(interaction, subtitle = {"channel" : "The selected channel does not support text messages."})
+        return
+
     try:
         try:
-            target_message = await channel.fetch_message(int(message_id))
+            target_message = await target_channel.fetch_message(int(message_id))
         except (discord.NotFound, ValueError, discord.HTTPException):
             await send_bad_argument(interaction, subtitle = {"message-id" : "The message provided does not exist, I lack permissions to access it, or it is not a valid ID."})
             return
 
         if target_message.author != interaction.client.user:
-            await send_bad_argument(interaction, subtitle = {"message-id" : "This messages was not sent by me, so I shouldn't delete it."})
+            await send_bad_argument(interaction, subtitle = {"message-id" : "The message provided was not sent by me, so I shouldn't delete it."})
             return
 
         await target_message.delete()
@@ -32,3 +39,4 @@ async def run_bo_messages_delete(
     except discord.Forbidden:
         _ = await send_unknown_error(interaction)
         return
+
